remove warning
remove warning

+++ Openholo/src/complex.h
@@ -107,7 +107,7 @@ namespace oph {
 			if (std::is_same<double, T>::value)
 				return atan2(_Val[_IM], _Val[_RE]);
 			else if (std::is_same<float, T>::value)
-				return atan2f(_Val[_IM], _Val[_RE]);
+				return atan2f((float)_Val[_IM], (float)_Val[_RE]);
 		}
 
 		Complex<T>& exp() {
